fix: grid view 기능이 동작하지 않던 현상 수정
pages와 newsdata 정보를 공유하고 있어야해 grid view에서 button에 event를
등록하는 로직을 main으로 가져옴
